rforms-engine/requirements.txt
# FastAPI and web server
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Database clients
psycopg>=3.1.18
psycopg-pool>=3.2.0
motor>=3.3.2
pymongo>=4.6.1
redis>=5.0.1
qdrant-client>=1.7.0
asyncpg==0.29.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2

# AI and vector operations
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
anthropic>=0.16.0  # For Claude API
openai>=1.12.0     # For OpenAI API

# Testing
pytest>=7.4.0
httpx>=0.27.0

# Utility
tenacity>=8.2.3
pydantic-settings>=2.1.0
loguru>=0.7.2
